Update vutil.tin
Decided to use & and . as global variables. This will result in clearer expectation for behavior.
Also, using "ref" variable for reference array instead of $@. This cleans up the output of RefSub.
Eliminated "numeric" type. The self-operation notation is sufficient. I want to have as little notation as possible.
The "lock" feature is now used in the GetEvalValue method to lock the "." variable. This ensures that behavior is made clear to the user.